+++ changes.md
@@ -1,7 +1,58 @@
 # PhysiCell: an Open Source Physics-Based Cell Simulator for 3-D Multicellular Systems
-**Versions:** 1.13.0 - 
+**Versions:** 1.14.0 - 
 
-**Release dates:** 29 July 2023 - 
+**Release dates:** ?? June 2024 - 
+* 1.14.0 : ?? June 2024
+
+changes: 
+* Add `Cell_Integrity` to `Phenotype`
+* rename `Phenotype.Interactions.damage_rate` to `Phenotype.Interactions.attack_damage_rate`
+* move `Cell.Cell_State.damage` to `Phenotype.Cell_Integrity.damage` 
+* added standardized (self-)damage and self repair to `Phenotype.Cell_Integrity` as part of built-in phenotype 
+
+
+updating ... phagocytosis models 
+updating .. effector atack models 
+
+* new (or modified) signals: 
+damage (now in cell integrity)
+damage delivered 
+attacking (as a boolean)
+
+* deprecated signals: 
+
+* new (or modified) behaviors: 
+phagocytose apoptotic cells
+phagocytose necrotic cells
+phagocytose other dead cells
+attack damage rate (renamed)
+damage rate (moved)
+damage repair rate 
+
+attack duration
+
+* deprecated behaviors: 
+phagocytose dead cells 
+
+rules: 
+give a warning on dead cell phago rules?
+
+
+
+
+to do: 
+* How to communicate to attacking cells that their target is dead. Fix for now: 
+
+
+to do: 
+* seprate necrotic and apoptotic phagocytosis rates
+* new signals: necrotic, apoptotic neighbors (as well as dead neighbors)
+
+## .. need to fill the rest in here. 
+
+
+**Versions:** 1.13.0 - 1.13.1 
+**Release dates:** 29 July 2023 - ?? June 2024 
 * 1.13.0 : 29 July 2023
 * 1.13.1 : 6 August 2023